(zsh i3) Extract monitor layout setup to a script
Now it is possible to setup the same bindings for layout settings in others
Desktop Environments. (sometimes I need to use xfce4).

In order to be able to run the new custom script the way how the i3
configuration file is created and updated had to change. Now I edit a
template file ($X_DOTFILES/i3/config) that is parsed in order to
generate the real config file (saved in ~/.config/i3/config)